+++ src/classes/Bootstrap.php
@@ -16,8 +16,11 @@ use Logeecom\Infrastructure\Serializer\Serializer;
 use Logeecom\Infrastructure\ServiceRegister;
 use Logeecom\Infrastructure\TaskExecution\Process;
 use Logeecom\Infrastructure\TaskExecution\QueueItem;
+use Packlink\Brands\Packlink\PacklinkConfigurationService;
 use Packlink\BusinessLogic\BootstrapComponent;
+use Packlink\BusinessLogic\Brand\BrandConfigurationService;
 use Packlink\BusinessLogic\Configuration;
+use Packlink\BusinessLogic\FileResolver\FileResolverService;
 use Packlink\BusinessLogic\Order\Interfaces\ShopOrderService as ShopOrderServiceInterface;
 use Packlink\BusinessLogic\OrderShipmentDetails\Models\OrderShipmentDetails;
 use Packlink\BusinessLogic\Scheduler\Models\Schedule;
@@ -28,12 +31,14 @@ use Packlink\PrestaShop\Classes\BusinessLogicServices\CarrierService;
 use Packlink\PrestaShop\Classes\BusinessLogicServices\ConfigurationService;
 use Packlink\PrestaShop\Classes\BusinessLogicServices\RegistrationInfoService;
 use Packlink\PrestaShop\Classes\BusinessLogicServices\ShopOrderService;
+use Packlink\PrestaShop\Classes\BusinessLogicServices\SystemInfoService;
 use Packlink\PrestaShop\Classes\Entities\CarrierServiceMapping;
 use Packlink\PrestaShop\Classes\Entities\CartCarrierDropOffMapping;
 use Packlink\PrestaShop\Classes\InfrastructureServices\LoggerService;
 use Packlink\PrestaShop\Classes\Repositories\BaseRepository;
 use Packlink\PrestaShop\Classes\Repositories\QueueItemRepository;
 use Packlink\BusinessLogic\Registration\RegistrationInfoService as RegistrationInfoServiceInterface;
+use Packlink\BusinessLogic\SystemInformation\SystemInfoService as SystemInfoInterface;
 
 /**
  * Class Bootstrap
@@ -71,6 +76,13 @@ class Bootstrap extends BootstrapComponent
         );
 
         ServiceRegister::registerService(
+            BrandConfigurationService::CLASS_NAME,
+            function () {
+                return new PacklinkConfigurationService();
+            }
+        );
+
+        ServiceRegister::registerService(
             ShopShippingMethodService::CLASS_NAME,
             function () {
                 return new CarrierService();
@@ -95,6 +107,23 @@ class Bootstrap extends BootstrapComponent
             RegistrationInfoServiceInterface::CLASS_NAME,
             function () {
                 return new RegistrationInfoService();
+            }
+        );
+
+        ServiceRegister::registerService(
+            SystemInfoInterface::CLASS_NAME,
+            function () {
+                return new SystemInfoService();
+            }
+        );
+
+        ServiceRegister::registerService(
+            FileResolverService::CLASS_NAME,
+            function () {
+                return new FileResolverService(array(
+                    dirname(__FILE__) . '/../views/brand/countries',
+                    dirname(__FILE__) . '/../views/countries',
+                ));
             }
         );
     }